+++ tests/register.spec.ts
@@ -1,25 +1,13 @@
-import { RegisterUser } from '../src/models/user.models';
+import { randomUserData } from '../src/factories/user.factory';
 import { LoginPage } from '../src/pages/login.page';
 import { RegisterPage } from '../src/pages/register.page';
 import { WelcomePage } from '../src/pages/welcome.page';
-import { faker } from '@faker-js/faker/locale/en';
 import { expect, test } from '@playwright/test';
 
 test.describe('Verify register', () => {
   test('register with correct data and login @GAD-R03-01', async ({ page }) => {
     // Arrange
-    const registerUserData: RegisterUser = {
-      userFirstName: faker.person.firstName().replace(/[^A-Za-z]/g, ''),
-      userLastName: faker.person.lastName().replace(/[^A-Za-z]/g, ''),
-      userEmail: '',
-      userPassword: faker.internet.password(),
-    };
-
-    registerUserData.userEmail = faker.internet.email({
-      firstName: registerUserData.userFirstName,
-      lastName: registerUserData.userLastName,
-    });
-
+    const registerUserData = randomUserData();
     const registerPage = new RegisterPage(page);
 
     // Act
@@ -50,12 +38,8 @@ test.describe('Verify register', () => {
     page,
   }) => {
     // Arrange
-    const registerUserData: RegisterUser = {
-      userFirstName: faker.person.firstName().replace(/[^A-Za-z]/g, ''),
-      userLastName: faker.person.lastName().replace(/[^A-Za-z]/g, ''),
-      userEmail: '#%^',
-      userPassword: faker.internet.password(),
-    };
+    const registerUserData = randomUserData();
+    registerUserData.userEmail = '@$#%';
 
     const expectedErrorText = `Please provide a valid email address`;
     const registerPage = new RegisterPage(page);
@@ -73,18 +57,15 @@ test.describe('Verify register', () => {
   }) => {
     // Arrange
     const expectedErrorText = `This field is required`;
+    const registerUserData = randomUserData();
     const registerPage = new RegisterPage(page);
 
     // Act
     await registerPage.goto();
 
-    await registerPage.userFirstNameInput.fill(
-      faker.person.firstName().replace(/[^A-Za-z]/g, ''),
-    );
-    await registerPage.userLastNameInput.fill(
-      faker.person.lastName().replace(/[^A-Za-z]/g, ''),
-    );
-    await registerPage.userPasswordInput.fill(faker.internet.password());
+    await registerPage.userFirstNameInput.fill(registerUserData.userFirstName);
+    await registerPage.userLastNameInput.fill(registerUserData.userLastName);
+    await registerPage.userPasswordInput.fill(registerUserData.userPassword);
     await registerPage.registerButton.click();
 
     // Assert
